Updating folders + types.
